Moved helpers to respective files
