+++ path_finding/__init__.py
@@ -1,0 +1,1 @@
+from settings import *
